generic: 5.15: backport smempart parser fixup patch with EPROBE_DEFER error
Backport patch from kernel 5.15 that mute error on EPROBE_DEFER with
smempart parser.

This parser require the smem device to be probed first and currently it
may happen that mtd gets probed before the smem device causing an error
on the smempart parser. This error may be confusing and should be muted.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>
Signed-off-by: Hou Qiyao <ryanhou24@gmail.com>